refactor: utils 파일로 포맷팅 함수들 분리
